Move openssl_csr_info code to module_utils.
